+++ src/base58.cpp
@@ -22,7 +22,7 @@ void Base58::init(int* INDEXES, char* ALPHABET)
 }
 
 // Encodes the given bytes as a base58 string (no checksum is appended).
-string Base58::encode(ubyte* data, uint len)
+string Base58::encode(ubyte* data, size_t len)
 {
     if (len == 0)
     {
@@ -35,22 +35,22 @@ string Base58::encode(ubyte* data, uint len)
 
     // Count leading zeros.
     int zeros = 0;
-    while (zeros < (int)len && data[zeros] == 0)
+    while ((zeros < (int)len) && (data[zeros] == 0))
     {
         ++zeros;
     }
 
     // Convert base-256 digits to base-58 digits (plus conversion to ASCII characters)
     ubyte* input = new ubyte[len];
-    for (uint i = 0; i < len; i++) // since we modify it in-place
+    for (size_t i = 0; i < len; i++) // since we modify it in-place
     {
         input[i] = data[i];
     }
 
     char* encoded = new char[len * 3]; // upper bound
-    uint outputStart = len * 3;
+    size_t outputStart = len * 3;
 
-    for (uint inputStart = zeros; inputStart < len;)
+    for (size_t inputStart = zeros; inputStart < len;)
     {
         encoded[--outputStart] = ALPHABET[divmod(input, len, inputStart, 256, 58)];
 
@@ -76,7 +76,7 @@ string Base58::encode(ubyte* data, uint len)
 }
 
 // Decodes the given base58 string into the original data bytes.
-uint Base58::decode(string const& data, ubyte* result)
+size_t Base58::decode(string const& data, ubyte* result)
 {
     if (data.length() == 0)
     {
@@ -91,7 +91,7 @@ uint Base58::decode(string const& data, ubyte* result)
     // Convert the base58-encoded ASCII chars to a base58 byte sequence (base58 digits).
     ubyte* input58 = new ubyte[data.length()];
 
-    for (uint i = 0; i < data.length(); ++i)
+    for (size_t i = 0; i < data.length(); ++i)
     {
         char c = data[i];
         int digit = (int)c < 128 ? INDEXES[(int)c] : -1;
@@ -105,7 +105,7 @@ uint Base58::decode(string const& data, ubyte* result)
     }
 
     // Count leading zeros.
-    uint zeros = 0;
+    size_t zeros = 0;
     while (zeros < data.length() && input58[zeros] == 0)
     {
         ++zeros;
@@ -113,11 +113,11 @@ uint Base58::decode(string const& data, ubyte* result)
 
     // Convert base-58 digits to base-256 digits.
     ubyte* decoded = new ubyte[data.length()];
-    uint outputStart = (uint)data.length();
+    size_t outputStart = data.length();
 
-    for (uint inputStart = zeros; inputStart < data.length();)
+    for (size_t inputStart = zeros; inputStart < data.length();)
     {
-        decoded[--outputStart] = divmod(input58, (int)data.length(), inputStart, 58, 256);
+        decoded[--outputStart] = divmod(input58, data.length(), inputStart, 58, 256);
 
         if (input58[inputStart] == 0)
         {
@@ -132,13 +132,13 @@ uint Base58::decode(string const& data, ubyte* result)
     }
 
     // Return decoded data (including original number of leading zeros).
-    for (uint i = outputStart - zeros; i < data.length(); i++)
+    for (size_t i = outputStart - zeros; i < data.length(); i++)
     {
         result[i - outputStart - zeros] = decoded[i];
     }
     result[data.length() - outputStart - zeros] = 0;
 
-    return (uint)(data.length() - outputStart - zeros);
+    return data.length() - outputStart - zeros;
 }
 
 /*
@@ -153,12 +153,12 @@ BigInt Base58::decodeToBigInteger(string input)
     return BigInt(buf, len);
 }
 
-ubyte Base58::divmod(ubyte* number, int len, int firstDigit, int base, int divisor)
+ubyte Base58::divmod(ubyte* number, size_t len, int firstDigit, int base, int divisor)
 {
     // this is just long division which accounts for the base of the input digits
     int remainder = 0;
 
-    for (int i = firstDigit; i < len; i++)
+    for (size_t i = firstDigit; i < len; i++)
     {
         int digit = (int)number[i] & 0xFF;
         int temp = remainder * base + digit;
